retrieve holiday from back - refacto/split MissionDetails
