+++ common/utils/apiQueries.js
@@ -826,6 +826,7 @@ export const ADMIN_COMPANIES_QUERY = gql`
               id
               name
               submitterId
+              isHoliday
               validations {
                 submitterId
                 receptionTime
@@ -874,6 +875,7 @@ export const ADMIN_COMPANIES_QUERY = gql`
               submitterId
               deletedAt
               deletedBy
+              isHoliday
               validations {
                 submitterId
                 receptionTime
